requirements.txt
# Core dependencies for PROJECT VERTEX
# Install with: pip install -r requirements.txt

pygame==2.6.1
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7   # optional but recommended — speeds up GL calls
opencv-python==4.11.0.86
mediapipe==0.10.21
